// frontend/src/api/model/trayItemDto.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * TeamPrompts.Api | v1
 * OpenAPI spec version: 1.0.0
 */
import type { ResultKind } from './resultKind';

export interface TrayItemDto {
  resultId: string;
  sessionId: string;
  content: string;
  kind: unknown | null | ResultKind;
  promptName: string;
  model: string;
  createdAt: string;
}